src/requirements.txt
django
python-decouple
mysqlclient
coverage
codecov
sphinx
sphinx-rtd-theme
gunicorn
psycopg2
djangorestframework
